minor
I created some empty php files that we will implement. Account, settings etc. are duplicated for now for student and instructor because ıdk how can we change the navigation bar if we have only one php (navigation bar is dfferent for instructor and student) but we have also admin so this looks like a bad idea but here we go, if you have something in mind tell me please